Add ability to set a default brand for organisation

+++ app/controllers/organisation_brands_controller.rb
@@ -23,9 +23,29 @@ class OrganisationBrandsController < WebController
     authorize organisation, :can_manage_organisation_brands?
 
     organisation_brand = organisation.organisation_brands.find_by!(brand_id: params[:id])
+
+    if organisation.default_brand_id == organisation_brand.brand_id
+      return redirect_to organisation_path(organisation), status: :see_other
+    end
+
     organisation_brand.destroy!
 
     redirect_to organisation_path(organisation), success: t(".success", brand_name: organisation_brand.brand.name)
+  end
+
+  def update_default
+    authorize organisation, :can_manage_organisation_brands?
+
+    if params[:brand_id].present?
+      brand = organisation.brands.find(params[:brand_id])
+      organisation.update!(default_brand: brand)
+
+      redirect_to organisation_path(organisation), success: t(".success", brand_name: brand.name)
+    else
+      organisation.update!(default_brand: nil)
+
+      redirect_to organisation_path(organisation), success: t(".success_govuk")
+    end
   end
 
 private
